fix(webserver): flacky execution controller tests
